Fixed bugs, added "active project", static method for adding tasks
Fixed bugs, for my deleteProject module that didn't reassign ids when
we deleted one project from the array. I've fixed this with UUID4 that
generates random IDs to my tasks and projects..

In initializeButtons.js module in controller folder I've added
static method so that we can assign toggle class of active to our buttons,
Also, now we have activeProject module in our controller folder
so that we can now which project is currently "active"  and based on that
where to append projects.

Made some reorganization of my code in ui.js that is not yet final.
Also, added addTask method in ui.js module for adding tasks to tasks array.

In the html template file, I've added "project-button" class and
"data-id" of 0 to my span icon and "Inbox" button so that, Inbox project
is the default project.  I have assign ID of 0 to my inbox object
and I've used Object.freeze in  to make ID static on the default
project so that It's always 0.

Made some reorganization of my code in ui.js that is not yet final.