CSS changes, hamburger nav menu, collapsing tasks
Added collapsing tasks so, when the user click on the task-item it shows
the message detail of the task.

Changed the look and feel of the form for editing and creating tasks
now It's we are using modal (dialog) element for it.

Added hamburger navbar menu for smaller screens with nice animation
It's now compatible on smaller screens.

Made several changes to my layout and CSS.

Fixed bug, where if we were in "Today" or "This Week" pages
we couldn't edit the tasks, now It's fully working.