Fixed bugs, updated UI, added validation function, dark mode
- I fixed bug when the user tries to edit task from "This Week" or "Today"
page and they get an error.

- Also, I've added form validation functionality so that users can't
add projects that have no title or that have the same name..

- And for task I also added some sort of form validation, where user
can't create an empty task.

- Also, I fixed some thing with the layout overlapping issues,
I've enhanced user experience on smaller screens. Hamburger nav menu
now closes when the user clicks the button in navigation.

We now use modals for creation and editing of tasks.

- Made fixes with gaps, margins, paddings.
- Added dark mode for the users.
- Added footer element.
- Added comments to my code, in both javascript and css files.